Cleanup

+++ src/Voicemeeter.UI.Graphics.D3D12/include/Voicemeeter/UI/Graphics/Frame.hpp
@@ -163,8 +163,10 @@ namespace Voicemeeter {
 						_first = false;
 					} else {
 						RECT rect{
-							pop(floor(_invalidFrom[0])), pop(floor(_invalidFrom[1])),
-							pop(ceil(_invalidTo[0])), pop(ceil(_invalidTo[1]))
+							static_cast<LONG>(pop(floor(_invalidFrom[0]))),
+							static_cast<LONG>(pop(floor(_invalidFrom[1]))),
+							static_cast<LONG>(pop(ceil(_invalidTo[0]))),
+							static_cast<LONG>(pop(ceil(_invalidTo[1])))
 						};
 						DXGI_PRESENT_PARAMETERS params{
 							1, &rect,
